% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desktime.R
\name{desktime}
\alias{desktime}
\title{Start a DeskTime timer}
\source{
The \href{https://www.themuse.com/advice/the-rule-of-52-and-17-its-random-but-it-ups-your-productivity}{52/17 Rule}
by \href{https://www.themuse.com/author/julia-gifford}{Julia Gifford}.
}
\usage{
desktime(
  work_time = 52,
  break_time = 17,
  start_color = "green",
  end_color = "red",
  sound = "ping",
  units = "minutes"
)
}
\arguments{
\item{work_time}{\verb{[numeric(1)]}\cr
The length of time for each work session.
Defaults to 52 minutes.}

\item{break_time}{\verb{[numeric(1)]}\cr
The length of time for each break.
Defaults to 17 minutes.}

\item{start_color, end_color}{\verb{[character(1)]}\cr
Colors used to display the time.
The time is displayed in \code{start_color} when there is time remaining,
and in \code{end_color} when the timer has expired.
Passed to \code{\link[cli:make_ansi_style]{cli::make_ansi_style()}}.
Defaults to \code{"green"} and \code{"red"}, respectively.}

\item{sound}{\verb{[numeric(1)]} or \verb{[character(1)]}
The sound played by \code{\link[beepr:beep]{beepr::beep()}} when a timer ends.
See \code{\link[beepr:beep]{beepr::beep()}} for possible values.

If \code{\link{NULL}}, sound is disabled.

If a timer has expired and hasn't been ended,
the sound will play again each time one-fifth of the target time passes
(for example, if a session lasts 25 minutes,
a reminder sound will play every 5 minutes after the session ends).}

\item{units}{\verb{[character(1)]}\cr
Units used when interpreting numeric time period arguments.
Defaults to \code{"minutes"}.
Accepted units are
\code{"seconds"}, \code{"minutes"}, \code{"hours"}, \code{"days"}, \code{"weeks"}, \code{"fortnights"}, \code{"months"}, \code{"years"}, \code{"decades"}, \code{"centuries"}, or \code{"millennia"}
(or abbreviations thereof).}
}
\value{
Invisibly returns an \link[R6:R6Class]{R6} object of class \code{DeskTime}.
}
\description{
Creates an interactive timer in your R session.
I recommend you don't call this function in your current R session,
as it will block the console.
You may consider running it from the command line in a new terminal window with

\if{html}{\out{<div class="sourceCode">}}\preformatted{R -e "breaktime::desktime()"
}\if{html}{\out{</div>}}
}
