\name{fortuneTeller}
\alias{fortuneTeller}

\title{fortuneTeller
}
\description{
The function gives a fortune cookie style fortune.
}
\usage{
fortuneTeller()
}

\details{
There are 20 fortunes that are randomly given, some are funny some are motivational. There is no interaction required from the user beyond executing the function.

\value{
A random string is returned.
}
\references{
Fortunes are from https://joshmadison.com/2008/04/20/fortune-cookie-fortunes/ and https://quotes.yourdictionary.com/articles/funny-fortune-cookie-sayings.html
}
\author{
Wen Yan Kang
}

\seealso{
magic8, diceGame the other 2 "games" included in this package
}
\examples{
##execute without arguments
fortuneTeller()

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~misc }
