% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomodoro.R
\name{pomodoro}
\alias{pomodoro}
\title{Start a Pomodoro timer}
\source{
The \href{https://francescocirillo.com/pages/pomodoro-technique}{Pomodoro Technique}
by \href{https://francescocirillo.com/pages/francesco-cirillo}{Francesco Cirillo}.
}
\usage{
pomodoro(
  work_time = 25,
  short_break_time = work_time/5,
  long_break_time = short_break_time * long_break_frequency,
  long_break_frequency = 4,
  start_color = "green",
  end_color = "red",
  sound = "ping",
  units = "minutes"
)
}
\arguments{
\item{work_time}{\verb{[numeric(1)]}\cr
The length of time for each working session.
Defaults to 25 minutes.}

\item{short_break_time}{\verb{[numeric(1)]}\cr
The length of time for each short break.
Defaults to 5 minutes.}

\item{long_break_time}{\verb{[numeric(1)]}\cr
The length of time for each long break in minutes.
Defaults to 20 minutes.}

\item{long_break_frequency}{\verb{[numeric(1)]}\cr
The number of working sessions between each long break.
Defaults to 4.}

\item{start_color, end_color}{\verb{[character(1)]}\cr
Colors used to display the time.
The time is displayed in \code{start_color} when there is time remaining,
and in \code{end_color} when the timer has expired.
Passed to \code{\link[cli:make_ansi_style]{cli::make_ansi_style()}}.
Defaults to \code{"green"} and \code{"red"}, respectively.}

\item{sound}{\verb{[numeric(1)]} or \verb{[character(1)]}
The sound played by \code{\link[beepr:beep]{beepr::beep()}} when a timer ends.
See \code{\link[beepr:beep]{beepr::beep()}} for possible values.

If \code{\link{NULL}}, sound is disabled.

If a timer has expired and hasn't been ended,
the sound will play again each time one-fifth of the target time passes
(for example, if a session lasts 25 minutes,
a reminder sound will play every 5 minutes after the session ends).}

\item{units}{\verb{[character(1)]}\cr
Units used when interpreting numeric time period arguments.
Defaults to \code{"minutes"}.
Accepted units are
\code{"seconds"}, \code{"minutes"}, \code{"hours"}, \code{"days"}, \code{"weeks"}, \code{"fortnights"}, \code{"months"}, \code{"years"}, \code{"decades"}, \code{"centuries"}, or \code{"millennia"}
(or abbreviations thereof).}
}
\value{
Invisibly returns an \link[R6:R6Class]{R6} object of class \code{Pomodoro}.
}
\description{
Creates an interactive timer in your R session.
I recommend you don't call this function in your current R session,
as it will block the console.
You may consider running it from the command line in a new terminal window with

\if{html}{\out{<div class="sourceCode">}}\preformatted{R -e "breaktime::pomodoro()"
}\if{html}{\out{</div>}}
}
