% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\alias{pomodoro}
\title{Start a timer}
\source{
Inspired by the \href{https://medium.com/@UrgentPigeon/the-flowtime-technique-7685101bd191}{Flowtime Technique}
by \href{https://medium.com/@UrgentPigeon/the-flowtime-technique-7685101bd191}{Urgent Pigeon}
and the \href{https://francescocirillo.com/pages/pomodoro-technique}{Pomodoro Technique}
by \href{https://francescocirillo.com/pages/francesco-cirillo}{Francesco Cirillo}.
}
\usage{
timer(
  work_time = 25,
  break_time = NULL,
  long_break_time = NULL,
  start_color = "green",
  end_color = "red"
)

pomodoro(
  work_time = 25,
  break_time = 5,
  long_break_time = 20,
  start_color = "green",
  end_color = "red"
)
}
\arguments{
\item{work_time}{\verb{[numeric(1)]}\cr
The length of time for each working session in minutes.
Defaults to 25 minutes.}

\item{break_time}{\verb{[numeric(1)]}\cr
The length of time for each short break in minutes.
If \code{\link{NULL}}, the break time will be one-fifth of the length of the previous
working session.}

\item{long_break_time}{\verb{[numeric(1)]}\cr
The length of time for each long break in minutes.
Long breaks occur every fourth break.
If \code{\link{NULL}}, the long break time will be one-fifth of the length of the
previous working session, plus the length of the previous three short breaks.}

\item{start_color, end_color}{\verb{[character(1)]}\cr
Colors used to display the time.
The time is displayed in \code{start_color} when there is time remaining in the
current session, and in \code{end_color} when the current session has expired.
Color specifications are handled by \code{\link[cli:make_ansi_style]{cli::make_ansi_style()}}.}
}
\value{
Invisibly returns an \link[R6:R6Class]{R6} object of class \code{Timer}
}
\description{
\code{timer()} creates an interactive timer in your R session.
I recommend you don't use \code{timer()} in your current R session,
as it will block the console.
You may consider running it from the command line in a new terminal window with

\if{html}{\out{<div class="sourceCode">}}\preformatted{R -e "breaktime::timer()"
}\if{html}{\out{</div>}}

\code{pomodoro()} is an alias with defaults that match the recommended fixed times
of the \href{https://francescocirillo.com/pages/pomodoro-technique}{Pomodoro Technique}:
25 minute working sessions, 5 minute breaks and 20 minute long breaks.
}
