% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaktime-package.R
\docType{package}
\name{breaktime-package}
\alias{breaktime}
\alias{breaktime-package}
\title{breaktime: Interactive Work Timer}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/breaktime}
  \item Report bugs at \url{https://github.com/rossellhayes/breaktime/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal2}
